libdm: timestamp: add missing uint64_t cast in non-realtime path
The HAVE_REALTIME version of _timestamp_to_uint64() correctly casts
tv_sec to uint64_t before multiplying by NSEC_PER_SEC (line 46),
but the non-realtime (gettimeofday) version omits the cast.

On systems where time_t is 32-bit, the multiplication
tv_sec * NSEC_PER_SEC overflows the 32-bit intermediate result
before being added to the uint64_t accumulator. Add the same
(uint64_t) cast used in the realtime version to both tv_sec and
tv_usec multiplications.

+++ libdm/libdm-timestamp.c
@@ -91,8 +91,8 @@ static uint64_t _timestamp_to_uint64(struct dm_timestamp *ts)
 {
 	uint64_t stamp = 0;
 
-	stamp += ts->t.tv_sec * NSEC_PER_SEC;
-	stamp += ts->t.tv_usec * NSEC_PER_USEC;
+	stamp += (uint64_t) ts->t.tv_sec * NSEC_PER_SEC;
+	stamp += (uint64_t) ts->t.tv_usec * NSEC_PER_USEC;
 
 	return stamp;
 }
